launchpad-agent/requirements.txt
livekit-agents[deepgram,elevenlabs,anthropic,silero]>=1.5.0
livekit-api>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
pydantic>=2.0.0
